+++ src/Config/Option/Mercure/EnableMercureConfigOption.php
@@ -4,10 +4,10 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Installer\Config\Option\Mercure;
 
-use Shlinkio\Shlink\Installer\Config\Option\AbstractSwooleDependentConfigOption;
+use Shlinkio\Shlink\Installer\Config\Option\Server\AbstractAsyncRuntimeDependentConfigOption;
 use Symfony\Component\Console\Style\StyleInterface;
 
-class EnableMercureConfigOption extends AbstractSwooleDependentConfigOption
+class EnableMercureConfigOption extends AbstractAsyncRuntimeDependentConfigOption
 {
     public const ENV_VAR = 'MERCURE_ENABLED';