+++ src/Config/Option/Visit/VisitsWebhooksConfigOption.php
@@ -4,14 +4,14 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Installer\Config\Option\Visit;
 
-use Shlinkio\Shlink\Installer\Config\Option\AbstractSwooleDependentConfigOption;
+use Shlinkio\Shlink\Installer\Config\Option\Server\AbstractAsyncRuntimeDependentConfigOption;
 use Shlinkio\Shlink\Installer\Config\Util\ConfigOptionsValidatorsTrait;
 use Symfony\Component\Console\Style\StyleInterface;
 
 use function implode;
 
 /** @deprecated */
-class VisitsWebhooksConfigOption extends AbstractSwooleDependentConfigOption
+class VisitsWebhooksConfigOption extends AbstractAsyncRuntimeDependentConfigOption
 {
     use ConfigOptionsValidatorsTrait;
 
@@ -26,7 +26,7 @@ class VisitsWebhooksConfigOption extends AbstractSwooleDependentConfigOption
     {
         return implode(',', $io->ask(
             'Provide a comma-separated list of webhook URLs which will receive POST notifications when short URLs '
-            . 'receive visits (Ignore this if you are not serving shlink with swoole or openswoole)',
+            . 'receive visits.',
             null,
             [$this, 'splitAndValidateMultipleUrls'],
         ));